fix: resolve remaining format warnings across modules (#2001)

+++ src/projects/base/ovlibrary/data.cpp
@@ -132,11 +132,11 @@ namespace ov
 	{
 		if (offset >= 0)
 		{
-			OV_ASSERT(GetLength() >= static_cast<size_t>(offset), "Invalid offset: %jd (length: %zu)", offset, GetLength());
+			OV_ASSERT(GetLength() >= static_cast<size_t>(offset), "Invalid offset: %jd (length: %zu)", static_cast<intmax_t>(offset), GetLength());
 		}
 		else
 		{
-			OV_ASSERT(GetLength() > static_cast<size_t>(-1LL * offset), "Invalid offset: %jd (length: %zu)", -1LL * offset, GetLength());
+			OV_ASSERT(GetLength() > static_cast<size_t>(-1LL * offset), "Invalid offset: %jd (length: %zu)", static_cast<intmax_t>(-1LL * offset), GetLength());
 		}
 
 		return Subdata(offset, (offset >= 0) ? (GetLength() - offset) : (static_cast<size_t>(-1 * offset)));
